+++ packages/storage/src/providers/s3/apis/internal/list.ts
@@ -31,8 +31,9 @@ import { logger } from '../../../../utils';
 import { DEFAULT_DELIMITER, STORAGE_INPUT_PREFIX } from '../../utils/constants';
 import { CommonPrefix } from '../../utils/client/s3data/types';
 import { IntegrityError } from '../../../../errors/IntegrityError';
-import { ListAPIInput, ListAPIOutput } from '../../types/inputs';
+import { ListAPIInput } from '../../types/inputs';
 import { ListAdvancedAPIInput } from '../../../../internals/types/inputs';
+import { ListAPIOutput } from '../../types/outputs';
 
 const MAX_PAGE_SIZE = 1000;
