+++ packages/storage/src/providers/s3/types/outputs.ts
@@ -124,6 +124,15 @@ export type ListPaginateWithPathOutput =
 	};
 
 /**
+ * @internal convinience types.
+ */
+export type ListAPIOutput =
+	| ListAllOutput
+	| ListPaginateOutput
+	| ListAllWithPathOutput
+	| ListPaginateWithPathOutput;
+
+/**
  *  Output type with path for S3 copy API.
  * @deprecated Use {@link CopyWithPathOutput} instead.
  */
